UserStorageRefactoring: Removed userId and deviceId usage from Storage and related classes
Removed userId and deviceId properties from Storage interface and their implementation in SharedPreferencesStorage, and removed references to them in ApphudInternal and UserRepository. Now user IDs are taken only from ApphudUser, which simplifies storage logic and prevents data duplication. This is a preparatory stage for further refactoring of work with users.